+++ src/processor/elasticsearch/mapping_stats/lookup.rs
@@ -2,7 +2,10 @@
 // or more contributor license agreements. Licensed under the Elastic License 2.0;
 // you may not use this file except in compliance with the Elastic License 2.0.
 
-use super::{super::Lookup, IndexMapping, MappingStats, MappingSummary};
+use super::{
+    super::{Lookup, missing_source_error},
+    IndexMapping, MappingStats, MappingSummary,
+};
 use eyre::Result;
 use futures::stream::{BoxStream, StreamExt};
 
@@ -19,6 +22,8 @@ impl From<MappingStats> for Lookup<MappingSummary> {
 impl Lookup<MappingSummary> {
     pub async fn from_stream(mut stream: BoxStream<'static, Result<(String, IndexMapping)>>) -> Self {
         let mut lookup = Lookup::new();
+        let mut saw_missing_source = false;
+        let mut saw_error = false;
         while let Some(result) = stream.next().await {
             match result {
                 Ok((index_name, index_mapping)) => {
@@ -26,10 +31,22 @@ impl Lookup<MappingSummary> {
                 }
                 Err(e) => {
                     tracing::warn!("Error reading from mapping stats stream: {}", e);
+                    if missing_source_error(&e) {
+                        saw_missing_source = true;
+                    } else {
+                        saw_error = true;
+                    }
                 }
             }
         }
-        lookup
+
+        if lookup.is_empty() && saw_missing_source && !saw_error {
+            Lookup::missing()
+        } else if saw_error {
+            lookup
+        } else {
+            lookup.was_parsed()
+        }
     }
 }